requirements.txt
pandas
pymongo
python-dotenv
scikit-learn
prophet
matplotlib
xgboost
imbalanced-learn
streamlit
apache-airflow
yfinance
lxml
openpyxl